+++ CHANGELOG.md
@@ -60,7 +60,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * (ledger) [#1277](https://github.com/evmos/ethermint/pull/1277) Add Ledger preprocessing transaction hook for EIP-712-signed Cosmos payloads.
 * (rpc) [#1296](https://github.com/evmos/ethermint/pull/1296) Add RPC Backend unit tests.
 * (rpc) [#1352](https://github.com/evmos/ethermint/pull/1352) Make the grpc queries run concurrently, don't block the consensus state machine.
-* (cli) [#1360](https://github.com/evmos/ethermint/pull/1360) Introduce a new `query-only` flag, such that when enabled (grpc|jsonrpc|grpc,jsonrpc), will start the node in a query-only mode.
+* (cli) [#1360](https://github.com/evmos/ethermint/pull/1360) Introduce `json-rpc-only` and `grpc-only` flag, such that when enabled, will start the node in a query-only mode.
 
 ### Bug Fixes